backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.5
faiss-cpu>=1.7.4
python-multipart>=0.0.7
aiofiles>=23.2.1
httpx>=0.26.0
beautifulsoup4>=4.12.3
mypy>=1.8.0
black>=24.1.0
isort>=5.13.2
flake8>=7.0.0
pytest>=8.0.0
